requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
billiard==4.2.0
celery==5.4.0
cfgv==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
distlib==0.3.8
Django==4.2.14
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-debug-toolbar==4.3.0
django-extensions==3.2.3
django-lifecycle==1.2.4
django-redis==5.4.0
django-timezone-field==7.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
filelock==3.15.4
identify==2.6.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.4.0
nodeenv==1.9.1
numpy==2.0.1
pandas==2.2.2
platformdirs==4.2.2
pre-commit==3.3.2
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
redis==5.0.4
referencing==0.35.1
rpds-py==0.20.0
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
uuid==1.30
vine==5.1.0
virtualenv==20.26.3
wcwidth==0.2.13
